Use query string params for auth
The url-based http Basic authorization that was
used for authrorizing accesses to Sense terminal
is deprecated in HTTP specs and causes Safari
to show "Possible Phishing Site" warning when
attempting to open a terminal window.

The changes in this commit are to prevent this
warning to show up. Instead of using Basic
authorization like:

https://sense:{password}@tty-{console_id}.sensegrid.com

, we pass username and password as query string
parameters like:

https://tty-{console_id}.sensegrid.com?u=sense&p={password}

Note that the keys for username and password are
'u' and 'p', respectively.