Fix both copy-and-paste & deselecting
The original unofficial patch for copy-and-
paste issue across browsers was causing
deselecting issue when the user tries to
select/highlight some texts in the terminal.

This patch has been reapplied with some
modifications so that it does not try to
select the hidden input element if there
is already something selected.