Fix tab complete
YUI compressor was used for compiling JavaScript
codes of butterfly and the codes generated with
it appear to have some negative side-effects
that does not exist in unminified codes. In this
case, tab completion does not work properly with
codes minified with YUI compressor. Changed
to use Closure compiler that does not introduce
such side-effects.